+++ device/tt_versim_device.cpp
@@ -308,7 +308,7 @@ std::uint32_t tt_VersimDevice::get_num_dram_channels(std::uint32_t device_id) {
     return get_soc_descriptor(device_id) -> get_num_dram_channels();
 }
 
-std::uint32_t tt_VersimDevice::get_dram_channel_size(std::uint32_t device_id, std::uint32_t channel) {
+std::uint64_t tt_VersimDevice::get_dram_channel_size(std::uint32_t device_id, std::uint32_t channel) {
     return get_soc_descriptor(device_id) -> dram_bank_size; // Space per channel is identical for now
 }
